+++ src/playlist.c
@@ -216,7 +216,7 @@ Playlist_add_callback(Playlist * self, PyObject *args,
     pl_callbacks_table_add(self, to_add);
 #ifdef DEBUG
     fprintf(stderr, "[DEBUG]-playlist- adding callback (%p,%p) py(%p,%p)\n",
-            pl_callbacks, tramp, PyFunction_GetCode(tramp->callback),
+            pl_callbacks, tramp, tramp->callback,
             tramp->userdata);
 #endif
     sp_playlist_add_callbacks(self->_playlist, pl_callbacks, tramp);
@@ -707,7 +707,7 @@ Playlist_remove_callback(Playlist * self, PyObject *args)
         return NULL;
 #ifdef DEBUG
     fprintf(stderr, "[DEBUG]-playlist- looking for callback py(%p,%p)\n",
-            PyFunction_GetCode(callback), userdata);
+            callback, userdata);
 #endif
     pl_callback = pl_callbacks_table_remove(self, callback, userdata);
     if (!pl_callback) {